LC-14 - Fixed unit test issue caused by App Insights filter.
