+++ src/test/java/uk.gov.cshr/controller/AuthenticationControllerTest.java
@@ -1,7 +1,5 @@
 package uk.gov.cshr.controller;
 
-import com.microsoft.applicationinsights.web.internal.WebRequestTrackingFilter;
-import org.apache.commons.lang3.reflect.FieldUtils;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -15,12 +13,7 @@ import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 import uk.gov.cshr.Application;
-
-import javax.servlet.Filter;
-import javax.servlet.FilterConfig;
-import javax.servlet.ServletContext;
-
-import java.util.Enumeration;
+import uk.gov.cshr.utils.MockMVCFilterOverrider;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
@@ -36,33 +29,8 @@ public class AuthenticationControllerTest {
     private MockMvc mockMvc;
 
     @Before
-    public void overrideFilters() throws IllegalAccessException {
-        Filter[] filters = (Filter[]) FieldUtils.readField(mockMvc, "filters", true);
-        for (int i = 0; i < filters.length; i++) {
-            if (filters[i].getClass().getSimpleName().equals("PatternMappingFilterProxy")) {
-                WebRequestTrackingFilter filter = (WebRequestTrackingFilter) FieldUtils.readField(filters[i],
-                        "delegate", true);
-                FieldUtils.writeField(filter, "appName", "test", true);
-                filter.init(new FilterConfig() {
-                    @Override
-                    public String getFilterName() {
-                        return null;
-                    }
-                    @Override
-                    public ServletContext getServletContext() {
-                        return null;
-                    }
-                    @Override
-                    public String getInitParameter(String name) {
-                        return null;
-                    }
-                    @Override
-                    public Enumeration<String> getInitParameterNames() {
-                        return null;
-                    }
-                });
-            }
-        }
+    public void overridePatternMappingFilterProxyFilter() throws IllegalAccessException {
+        MockMVCFilterOverrider.overrideFilterOf(mockMvc, "PatternMappingFilterProxy" );
     }
 
     @Test
